src/requirements.txt
numpy
selenium
pandas
bs4
psycopg2-binary
jinja2==3.0.1
jinjasql
odfpy
